Refactor corpus -> dictionary -> int (vocab size); add more types.
Getting a firmer hold on how the data is preprocessed, and how to eventually deal with larger datasets.

Fortunately, in the data.Corpus passed around, only the data.Dictionary within was used. Then, it turns out only the len() of the dictionary was used, which is the vocab size. This determines the size of the encoder and decoder matrices. Furthermore, the datasets returned are just --- I think --- 2D LongTensors where the data has been reshaped into 'batch size' columns.

Also, I belive that there is no vocab size limiting going on, which would explain why even a subset of larger data sets were crashing when only part of the dataset was on the GPU: the encoder and decoder matrices were massive. Will have to double-check this. It could be that limiting the vocab is enough.